% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr.mindfulness-package.R
\docType{package}
\name{sqlr.mindfulness-package}
\alias{sqlr.mindfulness}
\alias{sqlr.mindfulness-package}
\title{sqlr.mindfulness: R Package for the Sleep and Mindfulness Project}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

R package for the project: sleep and mindfulness: 
      a systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipsousp.github.io/sqlr.mindfulness/}
  \item \url{https://github.com/gipsousp/sqlr.mindfulness/}
  \item Report bugs at \url{https://github.com/gipsousp/sqlr.mindfulness/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Leticia Viana Sales \email{leticiamind@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]
  \item Ana Amelia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-4976-2623}{ORCID}) [scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder, copyright holder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
